project V1.5

+++ BL/BO/LineStation.cs
@@ -7,7 +7,7 @@ namespace BL.BO
     class LineStation : Station
     {
         public int LineId { get; set; }
-        public int DistanceToNextStation { get; set; }
+        public double DistanceToNextStation { get; set; }
         public TimeSpan TimeFromPrevStation { get; set; }
     }
 }
